+++ tests/utils/test_file_watcher.py
@@ -22,12 +22,16 @@ async def test_file_watcher(tmp_path: pathlib.Path) -> None:
     number_of_writes = 0
     expected_number_of_writes = 3
 
-    select = Select(timer=Timer(0.1), file_watcher=file_watcher)
-    while await select.ready():
-        if msg := select.timer:
-            filename.write_text(f"{msg.inner}")
-        elif msg := select.file_watcher:
-            assert msg.inner == filename
+    timer = Timer(0.1)
+
+    select = Select(timer, file_watcher)
+    async for ready_set in select.ready():
+        if timer in ready_set:
+            msg = timer.consume()
+            filename.write_text(f"{msg}")
+        if file_watcher in ready_set:
+            fname = file_watcher.consume()
+            assert fname == filename
             number_of_writes += 1
             # After receiving a write 3 times, unsubscribe from the writes channel
             if number_of_writes == expected_number_of_writes:
@@ -48,16 +52,22 @@ async def test_file_watcher_change_types(tmp_path: pathlib.Path) -> None:
         paths=[str(tmp_path)], event_types={FileWatcher.EventType.DELETE}
     )
 
-    select = Select(
-        write_timer=Timer(0.1), deletion_timer=Timer(0.5), watcher=file_watcher
-    )
+    write_timer = Timer(0.1)
+    deletion_timer = Timer(0.5)
+    watcher = file_watcher
+
+    select = Select(write_timer, deletion_timer, watcher)
     number_of_receives = 0
-    while await select.ready():
-        if msg := select.write_timer:
-            filename.write_text(f"{msg.inner}")
-        elif _ := select.deletion_timer:
+    async for ready_set in select.ready():
+        if write_timer in ready_set:
+            msg = write_timer.consume()
+            filename.write_text(f"{msg}")
+        if deletion_timer in ready_set:
+            _ = deletion_timer.consume()  # We need to consume the message
             os.remove(filename)
-        elif _ := select.watcher:
+        if watcher in ready_set:
+            fname = watcher.consume()
+            assert fname == filename
             number_of_receives += 1
             break
     assert number_of_receives == 1